fix: category hierarchy consolidation, score precision, FK integrity, misc validation
Fifth batch of backlog fixes from docs/backlog.md:

- Consolidated category hierarchy on CategoryRelationDB: dropped the unused
  CategoryDB.parent_category_id column (always null in practice — the
  frontend never exposed it), added a BFS cycle guard that rejects any
  parent_child relation (direct or transitive) that would close a loop
  while still allowing legitimate multi-parent taxonomies (Content Q5).
  Also fixed a real, separate pre-existing bug found while verifying this:
  GET /content/categories and GET /content/category-relations were
  completely unreachable (always 422) because GET /{content_id} was
  registered first and swallowed both paths — reordered route registration.
- DecisionResolutionDB.score changed from Integer to Float, dropping the
  int() cast that silently truncated weighted scores (Decision Q6).
- Added a mutual-exclusivity guard (DB CHECK + service-layer validator) so
  ModuleInstanceDB can never have both content_record_id and
  decision_slot_id set at once — implemented as "never both" rather than
  literally "exactly one required", since real seed data (a module_data-only
  cta module) legitimately leaves both null (Campaigns Q2).
- DeliveryExecutionDB.recipient_id is now a real FK to RecipientDB.id
  instead of storing the external_id string duplicately — removed the
  external_id lookup/translation from apply_event_to_preferences and
  send_send_instance entirely. DeliveryProvider.send()'s param renamed
  recipient_id -> recipient_email since the provider layer still needs an
  actual address (Insight Q5).
- Reject a category-assignment score of exactly zero at write time
  (Insight Q4).
- Fixed the apply_event_to_preferences dedupe key to include event_id, so
  two distinct legitimate engagements (e.g. clicks on two different links)
  no longer get conflated as duplicates (Insight Q3).
- OverrideEventDB.override_type is now a proper enum (content/segment) with
  a matching DB CHECK constraint (Overrides Q5).

Also fixes .claude/launch.json, which pointed uvicorn at the wrong module
path (app.main instead of main) and could never have started the server.

All items verified live end-to-end; full test suite (10/10) passes.

Updates docs/backlog.md: moves the 7 resolved items to Done.

+++ backend/app/campaigns/models.py
@@ -83,7 +83,7 @@ class DecisionResolution(BaseModel):
     content_record_id: int
     content_version_id: int | None = None
     reason: str | None = None
-    score: int | None = None
+    score: float | None = None
     created_at: datetime
 
 
@@ -92,4 +92,4 @@ class DecisionResolutionCreate(BaseModel):
     content_record_id: int
     content_version_id: int | None = None
     reason: str | None = None
-    score: int | None = None
+    score: float | None = None
